+++ tests/testthat/test-class_resources_qs.R
@@ -1,15 +1,15 @@
 tar_test("create tar_resources_qs object", {
-  x <- resources_qs_init(preset = "high")
+  x <- resources_qs_init()
   expect_silent(resources_validate(x))
 })
 
 tar_test("prohibit empty tar_resources_qs object", {
-  x <- resources_qs_init(preset = NULL)
+  x <- resources_qs_init(compress_level = NULL)
   expect_error(resources_validate(x), class = "tar_condition_validate")
 })
 
 tar_test("print tar_resources_qs object", {
-  x <- resources_qs_init(preset = "high")
+  x <- resources_qs_init()
   out <- utils::capture.output(print(x))
   expect_true(any(grepl("tar_resources_qs", out)))
 })