+++ tests/testthat/test-tar_resources_qs.R
@@ -3,14 +3,14 @@ tar_test("tar_resources_qs()", {
   expect_silent(resources_validate(out))
 })
 
-tar_test("tar_resources_qs() default preset", {
+tar_test("tar_resources_qs() non-default compression level", {
   tar_option_set(
     resources = tar_resources(
       qs = tar_resources_qs(
-        preset = "low"
+        compress_level = 4L
       )
     )
   )
   out <- tar_resources_qs()
-  expect_equal(out$preset, "low")
+  expect_equal(out$compress_level, 4L)
 })